+++ src/Presentation/SmartStore.Web/Infrastructure/DefaultWidgetSelector.cs
@@ -98,6 +98,7 @@ namespace SmartStore.Web.Infrastructure
 						{
 							Id = t.Id,
 							Bordered = t.WidgetBordered,
+							WrapContent = !t.WidgetWrapContent.HasValue || t.WidgetWrapContent.Value,
 							ShowTitle = t.WidgetShowTitle,
 							SystemName = t.SystemName.SanitizeHtmlId(),
 							Title = t.GetLocalized(x => t.Title),
@@ -137,6 +138,7 @@ namespace SmartStore.Web.Infrastructure
 									{ 
 										Id = widget.Id,
 										SystemName = widget.SystemName,
+										WrapContent = widget.WrapContent,
 										ShowTitle = widget.ShowTitle,
 										IsBordered = widget.Bordered,
 										Title = widget.Title,
@@ -211,6 +213,7 @@ namespace SmartStore.Web.Infrastructure
 			public int Id { get; set; }
 			public string[] WidgetZones { get; set; }
 			public string SystemName { get; set; }
+			public bool WrapContent { get; set; }
 			public bool ShowTitle { get; set; }
 			public bool Bordered { get; set; }
 			public string Title { get; set; }