+++ src/Presentation/SmartStore.Web/Models/Topics/TopicWidgetModel.cs
@@ -9,6 +9,7 @@ namespace SmartStore.Web.Models.Topics
         public string Title { get; set; }
         public string Html { get; set; }
 
+		public bool WrapContent { get; set; }
         public bool ShowTitle { get; set; }
         public bool IsBordered { get; set; }
     }